refactor: Phase 3 - promote DraftSelectionHandler → DraftController, rewrite index.php entrypoint
