+++ ibl5/tests/Mutation/controllers-handlers-pass1.txt
@@ -4,7 +4,7 @@ DepthChartEntryController
 DepthChartEntrySubmissionHandler
 DiscordWebhookHandler
 DraftHistoryApiHandler
-DraftSelectionHandler
+DraftController
 ETagHandler
 FileTypeHandler
 FranchiseRecordBookApiHandler
